+++ cogs/tibiawiki.py
@@ -368,13 +368,43 @@ class TibiaWiki:
 
         if monster["bestiary_class"] is not None:
             difficulties = {
-                "Trivial": "⭐",
-                "Easy": "⭐⭐",
-                "Medium": "⭐⭐⭐",
+                "Harmless": "▪▪▪▪",
+                "Trivial": "⭐▪▪▪",
+                "Easy": "⭐⭐▪▪",
+                "Medium": "⭐⭐⭐▪",
                 "Hard": "⭐⭐⭐⭐"
             }
+            occurrences = {
+                "Common": "🔹▪▪▪",
+                "Uncommon": "🔹🔹▪▪",
+                "Rare": "🔹🔹🔹▪",
+                "Very Rare": "🔹🔹🔹🔹",
+            }
+            kills = {
+                "Harmless": 25,
+                "Trivial": 250,
+                "Easy": 500,
+                "Medium": 1000,
+                "Hard": 2500
+            }
+            points = {
+                "Harmless": 1,
+                "Trivial": 5,
+                "Easy": 15,
+                "Medium": 25,
+                "Hard": 50
+            }
             difficulty = difficulties.get(monster["bestiary_level"], f"({monster['bestiary_level']})")
-            embed.add_field(name="Bestiary Class", value=f"{monster['bestiary_class']}\n{difficulty}")
+            occurrence = occurrences.get(monster["occurrence"], f"")
+            required_kills = kills[monster['bestiary_level']]
+            given_points =  points[monster['bestiary_level']]
+            if monster['occurrence'] == 'Very Rare':
+                required_kills = 5
+                given_points = max(points[monster['bestiary_level']]*2, 5)
+            kill_and_points = \
+                f"{required_kills:,} kills | {given_points}⚜️"
+            embed.add_field(name="Bestiary Class", value=f"{monster['bestiary_class']}\n{difficulty}\n{occurrence}"
+                                                         f"\n{kill_and_points}")
 
         # If monster drops no loot, we might as well show everything
         if long or not monster["loot"]: