+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - Increased /loot scanning speed.
 - Simultaneous loot scans are now user-wide, not global. Each user can only have one image scanned at a time.
 - `/version` now checks if the required minimum commit version of discord.py is being used.
-- Command names and aliases changes:
+- `/monster` now shows occurrence, kills required and charm points given.
+- Command name and aliases changes:
     - `/debug`: renamed to `/eval`.
     - `/help`: Alias `/commands` added.